refactor: move vulnerability detection constants to a dedicated file for better organization
