+++ plugins/wazuh-analysis/public/groups/threat-intelligence/apps/vulnerability-detection/application.tsx
@@ -4,11 +4,9 @@ import ReactDOM from 'react-dom';
 import { ThreatIntelligenceNavGroup } from '../..';
 import { Layout } from '../../../layout';
 import { createSideNavItems } from '../../../side-nav';
-import {
-  THREAT_INTELLIGENCE_TITLE,
-  VULNERABILITY_DETECTION_ID,
-} from '../../constants';
+import { THREAT_INTELLIGENCE_TITLE } from '../../constants';
 import { VulnerabilityDetectionApp } from './vulnerability-detection-app';
+import { VULNERABILITY_DETECTION_ID } from './constants';
 
 export const renderApp = async (params: AppMountParameters) => {
   const items = createSideNavItems({
