+++ plugins/wazuh-analysis/public/groups/threat-intelligence/apps/vulnerability-detection/constants.ts
@@ -1,0 +1,15 @@
+import { i18n } from '@osd/i18n';
+import { PLUGIN_ID } from '../../../../../common/constants';
+import { buildSubAppId } from '../../../../utils';
+import { THREAT_INTELLIGENCE_ID } from '../../constants';
+
+export const VULNERABILITY_DETECTION_ID = buildSubAppId(
+  THREAT_INTELLIGENCE_ID,
+  'vulnerability_detection',
+);
+export const VULNERABILITY_DETECTION_TITLE = i18n.translate(
+  `${PLUGIN_ID}.category.${VULNERABILITY_DETECTION_ID}`,
+  {
+    defaultMessage: 'Vulnerability Detection',
+  },
+);
