+++ plugins/wazuh-analysis/public/groups/threat-intelligence/apps/vulnerability-detection/vulnerability-detection-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { VULNERABILITY_DETECTION_TITLE } from '../../constants';
+import { VULNERABILITY_DETECTION_TITLE } from './constants';
 
 interface VulnerabilityDetectionAppProps {
   params: AppMountParameters;
